add .no-picom option

+++ install.sh
@@ -30,8 +30,11 @@ yes | cp -rf .bash_profile \
         $HOME/.local/bin
             mkdir -p $HOME/.config/nvim && cp .config/nvim/* $HOME/.config/nvim
             mkdir -p $HOME/.config/dunst && cp .config/dunst/dunstrc $HOME/.config/dunst
-            cp .config/picom.conf $HOME/.config/picom.conf
             mkdir -p $HOME/.ctags.d
             ln -s $HOME/.ctags $HOME/.ctags.d/default.ctags
             pip install black isort
             mkdir -p ~/.config/ulauncher/user-themes && git clone https://github.com/sudosubin/one-dark-ulauncher.git ~/.config/ulauncher/user-themes/one-dark-ulauncher
+
+            if [ ! -f .no-picom ]; then
+                cp .config/picom.conf $HOME/.config/picom.conf
+            fi
